add `SetSkipConcurrencyControl()` to transaction interface
The motivation for this is to allow disabling expensive per-key lock
tracking for _some_ keys inside a transaction.
Adding keys to the lock map in a transaction can be pretty expensive,
and avoiding this for some keys can have a positive impact on memory
usage and performance (lock map entries are quite heavy in terms of
memory usage, lock map stripes are protected by mutexes etc.).
`PessimisticTransaction` currently allows to set its
`skip_concurrency_control_` flag, but it must be set once for an entire
transaction and cannot be set for individual keys inside a transaction.
And it is only available in the `PessimisticTransaction` implementation,
which is a RocksDB-internal class that cannot be accessed from client
applications.

The suggested change in this PR would add a
`SetSkipConcurrencyControl(bool)` method to every transaction. The
default implementation will do nothing and always return
`Status::NotSupported()`.
For the `PessimisticTransaction` it is overridden so that the
`skip_concurrency_control_` flag of the transaction is modified. This
seems to be ok, as the flag is checked again upon every lock attempt.

There is no special override for `OptimisticTransaction`, simply because
I am not a user of `OptimisticTransaction` and don't know what the
correct behavior would be here.